+++ include/votca/xtp/dftengine.h
@@ -26,8 +26,9 @@
 #include <votca/xtp/convergenceacc.h>
 #include <votca/xtp/ecpaobasis.h>
 #include <votca/xtp/logger.h>
-#include <votca/xtp/numerical_integrations.h>
 #include <votca/xtp/staticsite.h>
+#include <votca/xtp/vxc_grid.h>
+#include <votca/xtp/vxc_potential.h>
 
 namespace votca {
 namespace xtp {
@@ -60,6 +61,8 @@ class DFTEngine {
  private:
   void Prepare(QMMolecule& mol);
 
+  Vxc_Potential<Vxc_Grid> SetupVxc(const QMMolecule& mol);
+
   Eigen::MatrixXd OrthogonalizeGuess(const Eigen::MatrixXd& GuessMOs) const;
   void PrintMOs(const Eigen::VectorXd& MOEnergies, Log::Level level);
   void CalcElDipole(const Orbitals& orb) const;
@@ -77,11 +80,11 @@ class DFTEngine {
                                         const Orbitals& extdensity) const;
 
   tools::EigenSystem IndependentElectronGuess(const Mat_p_Energy& H0) const;
-  tools::EigenSystem ModelPotentialGuess(const Mat_p_Energy& H0,
-                                         const QMMolecule& mol) const;
+  tools::EigenSystem ModelPotentialGuess(
+      const Mat_p_Energy& H0, const QMMolecule& mol,
+      const Vxc_Potential<Vxc_Grid>& vxcpotential) const;
 
   Eigen::MatrixXd AtomicGuess(const QMMolecule& mol) const;
-  std::string ReturnSmallGrid(const std::string& largegrid);
 
   Eigen::MatrixXd RunAtomicDFT_unrestricted(const QMAtom& uniqueAtom) const;
 
@@ -112,10 +115,6 @@ class DFTEngine {
 
   // numerical integration Vxc
   std::string _grid_name;
-  std::string _grid_name_small;
-  bool _use_small_grid;
-  NumericalIntegration _gridIntegration;
-  NumericalIntegration _gridIntegration_small;
 
   // AO Matrices
   AOOverlap _dftAOoverlap;
