+++ src/libxtp/grids/radial_euler_maclaurin_rule.cc
