+++ src/tests/CMakeLists.txt
@@ -22,7 +22,9 @@ if(ENABLE_TESTING)
   list(APPEND test_cases test_espfit)
   list(APPEND test_cases test_glink)
   list(APPEND test_cases test_hdf5)
-  list(APPEND test_cases test_numericalintegration)
+  list(APPEND test_cases test_densityintegration)
+list(APPEND test_cases test_vxc_potential)
+list(APPEND test_cases test_vxc_grid)
   list(APPEND test_cases test_orbitals)
   list(APPEND test_cases test_polarsite)
   list(APPEND test_cases test_ppm)
